+++ client/src/layout/CommonPageLayout/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import VerticalContainer from '../../common/VerticalContainer';
+import PageHeader from '../../common/PageHeader';
+import PageFooter from '../../common/PageFooter';
+
+function CommonPageLayout({ children }) {
+  return (
+    <VerticalContainer>
+      <PageHeader />
+      {children}
+      <PageFooter />
+    </VerticalContainer>
+  );
+}
+
+export default CommonPageLayout;
