+++ client/src/pages/BookInfoPage/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const CenterContainer = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: center;
+`;
+
+export { CenterContainer };
