+++ client/src/pages/OpenBoardPage/index.js
@@ -1,8 +1,14 @@
 import React from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
-import PageHeader from './common/PageHeader';
-import VerticalContainer from './common/VerticalContainer';
+import {
+  PageTitle,
+  Container,
+  Bottom,
+  Table,
+  TableHead,
+  UnorderedList }
+  from './styles';
+import CommonPageLayout from '../../layout/CommonPageLayout';
 
 const posts = [
   {
@@ -34,8 +40,7 @@ const posts = [
 function OpenBoardPage() {
   const navigate = useNavigate();
   return (
-    <VerticalContainer horizontal="center">
-      <PageHeader />
+    <CommonPageLayout>
       <PageTitle>
         <h1>자유게시판</h1>
         <p>자유롭게 글을 쓸 수 있는 공간입니다.</p>
@@ -86,7 +91,7 @@ function OpenBoardPage() {
           글쓰기
         </button>
       </Bottom>
-    </VerticalContainer>
+    </CommonPageLayout>
   );
 }
 
@@ -105,75 +110,5 @@ function Post({
 
   );
 }
-
-const PageTitle = styled.div`
-  width: 1100px;
-  margin: 1rem;
-
-  & h1 {
-    margin-top: 50px;
-    text-align: center;
-  }
-
-  & p {
-    text-align: center;
-  }
-`;
-
-const Container = styled.div`
-  width: 1100px;
-  margin: 1rem;
-  height: 50vh;
-  background-color: skyblue;
-`;
-
-const Table = styled.div`
-  width: 1100px;
-  margin: 0px;
-  background-color: pink;
-`;
-
-const TableHead = styled.div`
-  display: flex;
-
-  & div {
-    margin-right: 150px;
-    margin-left: 10px;
-    font-size: 20px;
-  }
-`;
-
-const UnorderedList = styled.ul`
-  display: flex;
-  list-style: none;
-  width: 500px;
-  margin-left: 500px;
-
-  & li {
-    margin: 5px;
-    font-size: 20px;
-  }
-`;
-
-const Bottom = styled.div`
-  display: flex;
-  width: 1100px;
-  align-items: center;
-
-  & button {
-    width: 100px;
-    background-color: white;
-    border: 3px solid gray;
-    border-radius: 20px;
-    padding: 6px;
-    font-size: 20px;
-    color: gray;
-  }
-
-  & button:hover {
-    background-color: slateblue;
-    color: white;
-  }
-`;
 
 export default OpenBoardPage;
